Mejoras y nuevas funcionalidades en la aplicación
Se realizaron ajustes en el manejo de logs en `MauiProgram.cs` y se eliminaron importaciones innecesarias en varios archivos. Se corrigieron formatos y métodos en `Estudiante.cs` y `Program.cs`. Se refactorizó `ClientesService.cs` para mejorar la configuración de `HttpClient`. Se añadieron nuevas clases y servicios para gestionar equipos y tipos de equipos, incluyendo `EquiposService` y `TiposEquiposService`. Se implementaron nuevas vistas `EquiposView` y `TiposEquiposView` con sus respectivos diseñadores y recursos. Se añadió un nuevo icono en `equipo-de-computadora.ico` y se mejoró la lógica de eliminación en `ClientesView.cs`.

+++ Desktop/Views/ClientesView.cs
@@ -111,17 +111,32 @@ namespace Desktop.Views
             //capturamos el cliente seleccionado en el DataGridView
             if (dataGridClientes.CurrentRow != null)
             {
-                clienteModificado = (Cliente)dataGridClientes.CurrentRow.DataBoundItem;
-                bool borradoOk = await clientesService.DeleteClienteAsync(clienteModificado.id);
-                if (!borradoOk)
+                var clienteAEliminar = (Cliente)dataGridClientes.CurrentRow.DataBoundItem;
+                // Preguntamos al usuario si está seguro de eliminar el cliente
+                var confirmResult = MessageBox.Show($"¿Está seguro de eliminar al cliente {clienteAEliminar.firstname} {clienteAEliminar.lastname}?", "Confirmar eliminación", MessageBoxButtons.YesNo);
+                if (confirmResult == DialogResult.Yes)
                 {
-                    tabControl.SelectedTab = tabPageLista;
+                    bool borradoOk = await clientesService.DeleteClienteAsync(clienteModificado.id);
+                    if (!borradoOk)
+                    {
+                        tabControl.SelectedTab = tabPageLista;
+                    }
+                    else
+                    {
+                        MessageBox.Show($"Cliente {clienteAEliminar.firstname} {clienteAEliminar.lastname} eliminado correctamente");
+                        LoadClientes();
+                    }
                 }
-                else
-                {
-                    MessageBox.Show("Cliente eliminado correctamente");
-                    LoadClientes();
-                }
+            }
+        }
+
+        private void txtBusqueda_KeyPress(object sender, KeyPressEventArgs e)
+        {
+            //chequeamos si la tecla presionada es Enter y pulsamos el botón de buscar
+            if (e.KeyChar == (char)Keys.Enter)
+            {
+                btnBuscar.PerformClick();
+                e.Handled = true; // Evita que el sonido de "ding" se reproduzca
             }
         }
     }
